+++ src/main/java/sopt/org/motivooServer/domain/auth/config/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package sopt.org.motivooServer.global.config.auth;
+package sopt.org.motivooServer.domain.auth.config;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
