+++ src/main/java/sopt/org/motivooServer/global/config/SecurityConfig.java
@@ -10,9 +10,9 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import sopt.org.motivooServer.global.config.auth.CustomAccessDeniedHandler;
-import sopt.org.motivooServer.global.config.auth.CustomJwtAuthenticationEntryPoint;
-import sopt.org.motivooServer.global.config.auth.JwtAuthenticationFilter;
+import sopt.org.motivooServer.domain.auth.config.CustomAccessDeniedHandler;
+import sopt.org.motivooServer.domain.auth.config.CustomJwtAuthenticationEntryPoint;
+import sopt.org.motivooServer.domain.auth.config.JwtAuthenticationFilter;
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
 @Configuration
